+++ CHANGELOG.md
@@ -5,12 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
 ### Added
 
 ### Changed
 
 ### Fixed
+
+
+## [1.17.3]
+### Added
+
+### Changed
+ - Hid last page button on Community Maps screen [#1208](https://github.com/PublicMapping/districtbuilder/pull/1208)
+
+### Fixed
+- Fixed adding regions to more than 1 worker [#1217](https://github.com/PublicMapping/districtbuilder/pull/1217)
 
 
 ## [1.17.2]
@@ -506,7 +516,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/publicmapping/districtbuilder/compare/1.17.2...HEAD
+[unreleased]: https://github.com/publicmapping/districtbuilder/compare/1.17.3...HEAD
+[1.17.3]: https://github.com/publicmapping/districtbuilder/compare/1.17.2...1.17.3
 [1.17.2]: https://github.com/publicmapping/districtbuilder/compare/1.17.1...1.17.2
 [1.17.1]: https://github.com/publicmapping/districtbuilder/compare/1.17.0...1.17.1
 [1.17.0]: https://github.com/publicmapping/districtbuilder/compare/1.16.1...1.17.0